Unit tests: made the TestProcessorEventSeeder standalone in source file
